requirements.txt
langchain>=0.1.0
openai==1.3.0
chromadb>=0.5.0
streamlit>=1.28.0
PyPDF2>=3.0.1
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
pydantic>=1.10.13
gradio>=4.0.0